+++ ReportingTests/DataMunging/TestForSkip/WeatherReportSkipDefinitions.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using FlatFile.FixedWidth.Interfaces;
+
+namespace DataMunging.Reporting.TestForSkip
+{
+    public class WeatherReportSkipDefinitions : ITestForSkip
+    {
+        public bool ShouldSkip(string row)
+        {
+            return row.StartsWith("  mo");
+        }
+    }
+}